Reorganize the code base
