+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright 2020-2021 QuantCo Inc, Christian Lorentzen 
+Copyright 2020-2021 QuantCo Inc, Christian Lorentzen
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
