+++ src/glum/_glm.py
@@ -1,47 +1,19 @@
-"""
-Generalized Linear Models with Exponential Dispersion Family
-
-Modified from code submitted as a PR to sklearn:
-https://github.com/scikit-learn/scikit-learn/pull/9405
-
-Original attribution from:
-https://github.com/scikit-learn/scikit-learn/pull/9405/files
-#diff-38e412190dc50455611b75cfcf2d002713dcf6d537a78b9a22cc6b1c164390d1
-'''
-Author: Christian Lorentzen <lorentzen.ch@googlemail.com>
-some parts and tricks stolen from other sklearn files.
-'''
-"""
-
-# License: BSD 3 clause
-import copy
 import re
 import sys
+import time
+import typing
 import warnings
-from collections.abc import Iterable, Mapping, Sequence
-from time import perf_counter
-from typing import Any, NamedTuple, Optional, Union, cast
+from collections.abc import Mapping, Sequence
+from typing import Any, Optional, Union
 
+import formulaic
 import numpy as np
+import packaging.version
 import pandas as pd
 import scipy.sparse as sps
-import scipy.sparse.linalg as splinalg
 import sklearn as skl
 import tabmat as tm
-from formulaic import Formula, FormulaSpec
-from formulaic.parser import DefaultFormulaParser
-from formulaic.utils.constraints import LinearConstraintParser
-from formulaic.utils.context import capture_context
-from packaging import version
 from scipy import linalg, sparse, stats
-from sklearn.base import BaseEstimator, RegressorMixin
-from sklearn.utils import check_array
-from sklearn.utils.validation import (
-    assert_all_finite,
-    check_consistent_length,
-    check_is_fitted,
-    column_or_1d,
-)
 
 from ._distribution import (
     BinomialDistribution,
@@ -53,8 +25,9 @@ from ._distribution import (
     NormalDistribution,
     PoissonDistribution,
     TweedieDistribution,
-    guess_intercept,
 )
+from ._formula import capture_context, parse_formula
+from ._intercept import guess_intercept
 from ._link import CloglogLink, IdentityLink, Link, LogitLink, LogLink, TweedieLink
 from ._solvers import (
     IRLSData,
@@ -64,656 +37,33 @@ from ._solvers import (
     _least_squares_solver,
     _trust_constr_solver,
 )
+from ._typing import ArrayLike, ShapedArrayLike, VectorLike, WaldTestResult
 from ._util import (
-    _add_missing_categories,
-    _align_df_categories,
-    _expand_categorical_penalties,
-    _is_contiguous,
-    _safe_toarray,
+    add_missing_categories,
+    align_df_categories,
+    expand_categorical_penalties,
+    is_contiguous,
+    safe_toarray,
+    standardize_warm_start,
+)
+from ._validation import (
+    check_array_tabmat_compliant,
+    check_offset,
+    check_weights,
+    check_X_y_tabmat_compliant,
 )
 
-if version.parse(skl.__version__).release < (1, 6):
+if packaging.version.parse(skl.__version__).release < (1, 6):
     keyword_finiteness = "force_all_finite"
-    validate_data = BaseEstimator._validate_data
+    validate_data = skl.base.BaseEstimator._validate_data
 else:
     keyword_finiteness = "ensure_all_finite"
     from sklearn.utils.validation import validate_data  # type: ignore
 
 _float_itemsize_to_dtype = {8: np.float64, 4: np.float32, 2: np.float16}
 
-VectorLike = Union[np.ndarray, pd.api.extensions.ExtensionArray, pd.Index, pd.Series]
 
-ArrayLike = Union[
-    list,
-    tm.MatrixBase,
-    tm.StandardizedMatrix,
-    pd.DataFrame,
-    sparse.spmatrix,
-    VectorLike,
-]
-
-ShapedArrayLike = Union[
-    tm.MatrixBase,
-    tm.StandardizedMatrix,
-    pd.DataFrame,
-    sparse.spmatrix,
-    VectorLike,
-]
-
-
-class WaldTestResult(NamedTuple):
-    test_statistic: float
-    p_value: float
-    df: int
-
-
-def check_array_tabmat_compliant(mat: ArrayLike, drop_first: bool = False, **kwargs):
-    to_copy = kwargs.get("copy", False)
-
-    if isinstance(mat, pd.DataFrame):
-        raise RuntimeError("DataFrames should have been converted by this point.")
-
-    if isinstance(mat, tm.SplitMatrix):
-        kwargs.update({"ensure_min_features": 0})
-        new_matrices = [
-            check_array_tabmat_compliant(m, drop_first=drop_first, **kwargs)
-            for m in mat.matrices
-        ]
-        new_indices = [elt.copy() for elt in mat.indices] if to_copy else mat.indices
-        return tm.SplitMatrix(new_matrices, new_indices)
-
-    if isinstance(mat, tm.CategoricalMatrix):
-        if to_copy:
-            return copy.copy(mat)
-        return mat
-
-    if isinstance(mat, tm.StandardizedMatrix):
-        return tm.StandardizedMatrix(
-            check_array_tabmat_compliant(mat.mat, drop_first=drop_first, **kwargs),
-            check_array(mat.shift, **kwargs),
-        )
-
-    original_type = type(mat)
-    if isinstance(mat, (tm.DenseMatrix, tm.SparseMatrix)):
-        res = check_array(mat.unpack(), **kwargs)
-    else:
-        res = check_array(mat, **kwargs)
-
-    if res is not mat and original_type in (tm.DenseMatrix, tm.SparseMatrix):
-        res = original_type(
-            res,
-            column_names=mat.column_names,  # type: ignore
-            term_names=mat.term_names,  # type: ignore
-        )
-
-    return res
-
-
-def check_X_y_tabmat_compliant(
-    X: ArrayLike, y: Union[VectorLike, sparse.spmatrix], **kwargs
-) -> tuple[Union[tm.MatrixBase, sparse.spmatrix, np.ndarray], np.ndarray]:
-    """
-    See the documentation for :func:`sklearn.utils.check_X_y`. This function
-    behaves identically for inputs that are not from the Matrix package and
-    fixes some parameters, such as ``'force_all_finite'``, to match the needs of
-    GLMs.
-
-    Returns
-    -------
-    X_converted : array-like
-        The converted and validated X.
-    y_converted : numpy.ndarray
-        The converted and validated y.
-    """
-    if y is None:
-        raise ValueError("y cannot be None")
-
-    y = column_or_1d(y, warn=True)
-    assert_all_finite(y)
-    if y.dtype.kind == "O":
-        y = y.astype(np.float64)
-
-    check_consistent_length(X, y)
-    X = check_array_tabmat_compliant(X, **kwargs)
-
-    return X, y
-
-
-def _check_weights(
-    sample_weight: Optional[Union[float, VectorLike]],
-    n_samples: int,
-    dtype,
-    force_all_finite: bool = True,
-) -> np.ndarray:
-    """Check that sample weights are non-negative and have the right shape."""
-    if sample_weight is None:
-        return np.ones(n_samples, dtype=dtype)
-    if np.isscalar(sample_weight):
-        if sample_weight <= 0:  # type: ignore
-            raise ValueError("Sample weights must be non-negative.")
-        return np.full(n_samples, sample_weight, dtype=dtype)
-
-    sample_weight = check_array(
-        sample_weight,
-        accept_sparse=False,
-        ensure_2d=False,
-        dtype=[np.float64, np.float32],
-        **{keyword_finiteness: force_all_finite},
-    )
-
-    if sample_weight.ndim > 1:  # type: ignore
-        raise ValueError("Sample weights must be 1D array or scalar.")
-    if sample_weight.shape[0] != n_samples:  # type: ignore
-        raise ValueError("Sample weights must have the same length as y.")
-    if np.any(sample_weight < 0):  # type: ignore
-        raise ValueError("Sample weights must be non-negative.")
-    if np.sum(sample_weight) == 0:  # type: ignore
-        raise ValueError("Sample weights must have at least one positive element.")
-
-    return sample_weight  # type: ignore
-
-
-def _check_offset(
-    offset: Optional[Union[VectorLike, float]], n_rows: int, dtype
-) -> Optional[np.ndarray]:
-    """
-    Unlike weights, if the offset is ``None``, it can stay ``None``, so we only
-    need to validate it when it is not.
-    """
-    if offset is None:
-        return None
-    if np.isscalar(offset):
-        return np.full(n_rows, offset)
-
-    offset = check_array(
-        offset,
-        accept_sparse=False,
-        ensure_2d=False,
-        dtype=dtype,
-        **{keyword_finiteness: True},
-    )
-
-    offset = cast(np.ndarray, offset)
-
-    if offset.ndim > 1:  # type: ignore
-        raise ValueError("Offsets must be 1D array or scalar.")
-    if offset.shape[0] != n_rows:  # type: ignore
-        raise ValueError("Offsets must have the same length as y.")
-
-    return offset
-
-
-def _parse_formula(
-    formula: FormulaSpec, include_intercept: bool = True
-) -> tuple[Optional[Formula], Formula]:
-    """
-    Parse and transform the formula for use in a GeneralizedLinearRegressor.
-
-    The left-hand side and right-hand side of the formula are separated. If an
-    intercept is present, it will be removed from the right-hand side, and a
-    boolean flag to indicate whether or not an intercept should be added to
-    the model will be returned.
-
-    Parameters
-    ----------
-    formula : formulaic.FormulaSpec
-        The formula to parse.
-    include_intercept: bool, default True
-        Whether to include an intercept column.
-
-    Returns
-    -------
-    tuple[formulaic.Formula, formulaic.Formula]
-        The left-hand side and right-hand sides of the formula.
-    """
-    if isinstance(formula, str):
-        parser = DefaultFormulaParser(include_intercept=include_intercept)
-        terms = parser.get_terms(formula)
-    elif isinstance(formula, Formula):
-        terms = formula
-    else:
-        raise TypeError("formula must be a string or Formula object.")
-
-    if hasattr(terms, "lhs"):
-        lhs_terms = terms.lhs
-        if len(lhs_terms) != 1:
-            raise ValueError(
-                "formula must have exactly one term on the left-hand side."
-            )
-        rhs_terms = terms.rhs
-    else:
-        lhs_terms = None
-        rhs_terms = terms
-
-    return lhs_terms, rhs_terms
-
-
-def check_bounds(
-    bounds: Optional[Union[float, VectorLike]], n_features: int, dtype
-) -> Optional[np.ndarray]:
-    """Check that the bounds have the right shape."""
-    if bounds is None:
-        return None
-    if np.isscalar(bounds):
-        return np.full(n_features, bounds, dtype=dtype)
-
-    bounds = check_array(
-        bounds,
-        accept_sparse=False,
-        ensure_2d=False,
-        dtype=dtype,
-        **{keyword_finiteness: False},
-    )
-
-    bounds = cast(np.ndarray, bounds)
-
-    if bounds.ndim > 1:  # type: ignore
-        raise ValueError("Bounds must be 1D array or scalar.")
-    if bounds.shape[0] != n_features:  # type: ignore
-        raise ValueError("Bounds must be the same length as X.shape[1].")
-
-    return bounds
-
-
-def check_inequality_constraints(
-    A_ineq: Optional[np.ndarray],
-    b_ineq: Optional[np.ndarray],
-    n_features: int,
-    dtype,
-) -> tuple[Union[None, np.ndarray], Union[None, np.ndarray]]:
-    """Check that the inequality constraints are well-defined."""
-    if A_ineq is None or b_ineq is None:
-        return None, None
-    else:
-        A_ineq = check_array(
-            A_ineq,
-            accept_sparse=False,
-            ensure_2d=True,
-            dtype=dtype,
-            copy=True,
-            **{keyword_finiteness: False},
-        )
-        b_ineq = check_array(
-            b_ineq,
-            accept_sparse=False,
-            ensure_2d=False,
-            dtype=dtype,
-            copy=True,
-            **{keyword_finiteness: False},
-        )
-        if A_ineq.shape[1] != n_features:  # type: ignore
-            raise ValueError("A_ineq must have same number of columns as X.")
-        if A_ineq.shape[0] != b_ineq.shape[0]:  # type: ignore
-            raise ValueError("A_ineq and b_ineq must have same number of rows.")
-        if b_ineq.ndim > 1:  # type: ignore
-            raise ValueError("b_ineq must be 1D array.")
-    return A_ineq, b_ineq
-
-
-def _standardize(
-    X: tm.MatrixBase,
-    sample_weight: np.ndarray,
-    center_predictors: bool,
-    estimate_as_if_scaled_model: bool,
-    lower_bounds: Optional[np.ndarray],
-    upper_bounds: Optional[np.ndarray],
-    A_ineq: Optional[np.ndarray],
-    P1: Union[np.ndarray, sparse.spmatrix],
-    P2: Union[np.ndarray, sparse.spmatrix],
-) -> tuple[
-    tm.StandardizedMatrix,
-    np.ndarray,
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Optional[np.ndarray],
-    Any,
-    Any,
-]:
-    """
-    Standardize the data matrix ``X`` and adjust the bounds and penalties to
-    match the standardized data matrix, so that standardizing does not affect
-    estimates.
-
-    This is only done for computational reasons and does not affect final
-    estimates or alter the input data. Columns are always scaled to have unit
-    standard deviation.
-
-    Bounds, inequality constraints and regularization coefficients are modified
-    appropriately so that the estimates remain unchanged compared to an
-    unstandardized problem.
-
-    Parameters
-    ----------
-    X : MatrixBase
-    sample_weight : numpy.ndarray
-    center_predictors : bool
-        If ``True``, adjust the data matrix so that columns have mean zero.
-    estimate_as_if_scaled_model : bool
-        If ``True``, estimates returned equal those from a model where
-        predictors have been standardized to have unit standard deviation, with
-        penalty unchanged. Note that, internally, for purely computational
-        reasons, we always scale predictors; whether estimates match a scaled
-        model depends on whether we modify the penalty. If ``False``, penalties
-        are rescaled to match the original scale, canceling out the effect of
-        rescaling X.
-    lower_bounds
-    upper_bounds
-    A_ineq
-    P1
-    P2
-    """
-    X, col_means, col_stds = X.standardize(sample_weight, center_predictors, True)
-
-    if col_stds is not None:
-        inv_col_stds = _one_over_var_inf_to_val(col_stds, 1.0)
-        # We copy the bounds when multiplying here so the we avoid
-        # side effects.
-        if lower_bounds is not None:
-            lower_bounds = lower_bounds / inv_col_stds
-        if upper_bounds is not None:
-            upper_bounds = upper_bounds / inv_col_stds
-        if A_ineq is not None:
-            A_ineq = A_ineq * inv_col_stds
-
-    if not estimate_as_if_scaled_model and col_stds is not None:
-        P1 *= inv_col_stds
-        if sparse.issparse(P2):
-            inv_col_stds_mat = sparse.diags(inv_col_stds)
-            P2 = inv_col_stds_mat @ P2 @ inv_col_stds_mat
-        elif P2.ndim == 1:
-            P2 *= inv_col_stds**2
-        else:
-            P2 = (inv_col_stds[:, None] * P2) * inv_col_stds[None, :]
-
-    return X, col_means, col_stds, lower_bounds, upper_bounds, A_ineq, P1, P2
-
-
-def _unstandardize(
-    col_means: np.ndarray,
-    col_stds: Optional[np.ndarray],
-    intercept: Union[float, np.ndarray],
-    coef: np.ndarray,
-) -> tuple[Union[float, np.ndarray], np.ndarray]:
-    if col_stds is None:
-        intercept -= np.squeeze(np.squeeze(col_means).dot(np.atleast_1d(coef).T))
-    else:
-        penalty_mult = _one_over_var_inf_to_val(col_stds, 1.0)
-        intercept -= np.squeeze(
-            np.squeeze(col_means * penalty_mult).dot(np.atleast_1d(coef).T)
-        )
-        coef *= penalty_mult
-    return intercept, coef
-
-
-def _one_over_var_inf_to_val(arr: np.ndarray, val: float) -> np.ndarray:
-    """
-    Return 1/arr unless the values are zeros.
-
-    If values are zeros, return val.
-    """
-    zeros = np.where(np.abs(arr) < 10 * np.sqrt(np.finfo(arr.dtype).eps))
-    with np.errstate(divide="ignore"):
-        one_over = 1 / arr
-    one_over[zeros] = val
-    return one_over
-
-
-def _standardize_warm_start(
-    coef: np.ndarray, col_means: np.ndarray, col_stds: Optional[np.ndarray]
-) -> None:
-    if col_stds is None:
-        coef[0] += np.squeeze(col_means).dot(coef[1:])
-    else:
-        coef[1:] *= col_stds
-        coef[0] += np.squeeze(col_means * _one_over_var_inf_to_val(col_stds, 1)).dot(
-            coef[1:]
-        )
-
-
-def get_family(
-    family: Union[str, ExponentialDispersionModel],
-) -> ExponentialDispersionModel:
-    if isinstance(family, ExponentialDispersionModel):
-        return family
-
-    name_to_dist = {
-        "binomial": BinomialDistribution(),
-        "gamma": GammaDistribution(),
-        "gaussian": NormalDistribution(),
-        "inverse.gaussian": InverseGaussianDistribution(),
-        "normal": NormalDistribution(),
-        "poisson": PoissonDistribution(),
-        "tweedie": TweedieDistribution(1.5),
-        "negative.binomial": NegativeBinomialDistribution(1.0),
-    }
-
-    if family in name_to_dist:
-        return name_to_dist[family]
-
-    custom_tweedie = re.search(r"tweedie\s?\((.+)\)", family)
-
-    if custom_tweedie:
-        return TweedieDistribution(float(custom_tweedie.group(1)))
-
-    custom_negative_binomial = re.search(r"negative.binomial\s?\((.+)\)", family)
-
-    if custom_negative_binomial:
-        return NegativeBinomialDistribution(float(custom_negative_binomial.group(1)))
-
-    raise ValueError(
-        "The family must be an instance of class ExponentialDispersionModel or an "
-        f"element of {sorted(name_to_dist.keys())}; got (family={family})."
-    )
-
-
-def get_link(link: Union[str, Link], family: ExponentialDispersionModel) -> Link:
-    """
-    For the Tweedie distribution, this code follows actuarial best practices
-    regarding link functions. Note that these links are sometimes not canonical:
-        - identity for normal (``p = 0``);
-        - no convention for ``p < 0``, so let's leave it as identity;
-        - log otherwise.
-    """
-    if isinstance(link, Link):
-        return link
-
-    if (link is None) or (link == "auto"):
-        if tweedie_representation := family.to_tweedie(safe=False):
-            if tweedie_representation.power <= 0:
-                return IdentityLink()
-            return LogLink()
-        if isinstance(family, GeneralizedHyperbolicSecant):
-            return IdentityLink()
-        if isinstance(family, BinomialDistribution):
-            return LogitLink()
-        if isinstance(family, NegativeBinomialDistribution):
-            return LogLink()
-        raise ValueError(
-            "No default link known for the specified distribution family. "
-            "Please set link manually, i.e. not to 'auto'. "
-            f"Got (link='auto', family={family.__class__.__name__})."
-        )
-
-    mapping = {
-        "cloglog": CloglogLink(),
-        "identity": IdentityLink(),
-        "log": LogLink(),
-        "logit": LogitLink(),
-        "tweedie": TweedieLink(1.5),
-    }
-
-    if link in mapping:
-        return mapping[link]
-    if custom_tweedie := re.search(r"tweedie\s?\((.+)\)", link):
-        return TweedieLink(float(custom_tweedie.group(1)))
-
-    raise ValueError(
-        "The link must be an instance of class Link or an element of "
-        "['auto', 'identity', 'log', 'logit', 'cloglog', 'tweedie']; "
-        f"got (link={link})."
-    )
-
-
-def setup_p1(
-    P1: Optional[Union[str, np.ndarray]],
-    X: Union[tm.MatrixBase, tm.StandardizedMatrix],
-    dtype,
-    alpha: float,
-    l1_ratio: float,
-) -> np.ndarray:
-    if not isinstance(X, (tm.MatrixBase, tm.StandardizedMatrix)):
-        raise TypeError
-
-    n_features = X.shape[1]
-
-    if isinstance(P1, str):
-        if P1 != "identity":
-            raise ValueError(f"P1 must be either 'identity' or an array; got {P1}.")
-        P1 = np.ones(n_features, dtype=dtype)
-    elif P1 is None:
-        P1 = np.ones(n_features, dtype=dtype)
-    else:
-        P1 = np.atleast_1d(P1)
-        try:
-            P1 = P1.astype(dtype, casting="safe", copy=False)  # type: ignore
-        except TypeError as e:
-            raise TypeError(
-                "The given P1 cannot be converted to a numeric array; "
-                f"got (P1.dtype={P1.dtype})."  # type: ignore
-            ) from e
-        if (P1.ndim != 1) or (P1.shape[0] != n_features):  # type: ignore
-            raise ValueError(
-                "P1 must be either 'identity' or a 1d array with the length of "
-                "X.shape[1] (either before or after categorical expansion); "
-                f"got (P1.shape[0]={P1.shape[0]})."  # type: ignore
-            )
-
-    # P1 and P2 are now for sure copies
-    P1 = alpha * l1_ratio * P1  # type: ignore
-    return cast(np.ndarray, P1).astype(dtype)
-
-
-def setup_p2(
-    P2: Optional[Union[str, np.ndarray, sparse.spmatrix]],
-    X: Union[tm.MatrixBase, tm.StandardizedMatrix],
-    stype,
-    dtype,
-    alpha: float,
-    l1_ratio: float,
-) -> Union[np.ndarray, sparse.spmatrix]:
-    if not isinstance(X, (tm.MatrixBase, tm.StandardizedMatrix)):
-        raise TypeError
-
-    n_features = X.shape[1]
-
-    def _setup_sparse_p2(P2):
-        return (sparse.dia_matrix((P2, 0), shape=(n_features, n_features))).tocsc()
-
-    if isinstance(P2, str):
-        if P2 != "identity":
-            raise ValueError(f"P2 must be either 'identity' or an array. Got {P2}.")
-        if sparse.issparse(X):  # if X is sparse, make P2 sparse, too
-            P2 = _setup_sparse_p2(np.ones(n_features, dtype=dtype))
-        else:
-            P2 = np.ones(n_features, dtype=dtype)
-    elif P2 is None:
-        if sparse.issparse(X):  # if X is sparse, make P2 sparse, too
-            P2 = _setup_sparse_p2(np.ones(n_features, dtype=dtype))
-        else:
-            P2 = np.ones(n_features, dtype=dtype)
-    else:
-        P2 = check_array(
-            P2, copy=True, accept_sparse=stype, dtype=dtype, ensure_2d=False
-        )
-        P2 = cast(np.ndarray, P2)
-        if P2.ndim == 1:
-            P2 = np.asarray(P2)
-            if P2.shape[0] != n_features:
-                raise ValueError(
-                    "P2 should be a 1d array of shape X.shape[1] (either before or "
-                    "after categorical expansion); "
-                    f"got (P2.shape={P2.shape})."
-                )
-            if sparse.issparse(X):
-                P2 = _setup_sparse_p2(P2)
-        elif P2.ndim == 2 and P2.shape[0] == P2.shape[1] and P2.shape[0] == n_features:
-            if sparse.issparse(X):
-                P2 = sparse.csc_matrix(P2)
-        else:
-            raise ValueError(
-                "P2 must be either None or an array of shape (n_features, n_features) "
-                f"with n_features=X.shape[1]; got (P2.shape={P2.shape}); "
-                f"needed ({n_features}, {n_features})."
-            )
-
-    # P1 and P2 are now for sure copies
-    P2 = alpha * (1 - l1_ratio) * P2
-    # one only ever needs the symmetrized L2 penalty matrix 1/2 (P2 + P2')
-    # reason: w' P2 w = (w' P2 w)', i.e. it is symmetric
-    if P2.ndim == 2:
-        if sparse.issparse(P2):
-            if sparse.isspmatrix_csc(P2):
-                P2 = 0.5 * (P2 + P2.transpose()).tocsc()
-            else:
-                P2 = 0.5 * (P2 + P2.transpose()).tocsr()
-        else:
-            P2 = 0.5 * (P2 + P2.T)
-    return P2
-
-
-def is_pos_semidef(p: Union[sparse.spmatrix, np.ndarray]) -> Union[bool, np.bool_]:
-    """
-    Checks for positive semidefiniteness of ``p`` if ``p`` is a matrix, or
-    ``diag(p)`` if a vector.
-
-    ``np.linalg.cholesky(P2)`` 'only' asserts positive definiteness; due to
-    numerical precision, we allow eigenvalues to be a tiny bit negative.
-    """
-    # 1d case
-    if p.ndim == 1 or p.shape[0] == 1:
-        any_negative = (p < 0).max() if sparse.isspmatrix(p) else (p < 0).any()
-        return not any_negative
-
-    # 2d case
-    # About -6e-7 for 32-bit, -1e-15 for 64-bit
-    epsneg = -10 * np.finfo(np.result_type(float, p.dtype)).epsneg
-
-    if sparse.issparse(p):
-        # Computing eigenvalues for sparse matrices is inefficient. If the matrix is
-        # not huge, convert to dense. Otherwise, calculate 10% of its eigenvalues.
-        p = cast(sparse.spmatrix, p)
-        if p.shape[0] < 2000:
-            eigenvalues = linalg.eigvalsh(p.toarray())
-        else:
-            n_evals_to_compuate = p.shape[0] // 10 + 1
-            sigma = -1000 * epsneg  # start searching near this value
-            which = "SA"  # find smallest algebraic eigenvalues first
-            eigenvalues = splinalg.eigsh(
-                p,
-                k=n_evals_to_compuate,
-                sigma=sigma,
-                which=which,
-                return_eigenvectors=False,
-            )
-    else:  # dense
-        eigenvalues = linalg.eigvalsh(p)
-
-    return np.all(eigenvalues >= epsneg)
-
-
-def _group_sum(groups: np.ndarray, data: tm.MatrixBase):
-    """Sum over groups."""
-    ngroups = len(np.unique(groups))
-    out = np.empty((ngroups, data.shape[1]))
-    eye_n = sps.eye(ngroups, format="csc")[:, groups]
-    for i in range(data.shape[1]):
-        out[:, i] = _safe_toarray(eye_n @ data.getcol(i).unpack()).ravel()
-    return out
-
-
-# TODO: abc
-class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
+class GeneralizedLinearRegressorBase(skl.base.RegressorMixin, skl.base.BaseEstimator):
     """
     Base class for :class:`GeneralizedLinearRegressor` and
     :class:`GeneralizedLinearRegressorCV`.
@@ -754,7 +104,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         drop_first: bool = False,
         robust: bool = True,
         expected_information: bool = False,
-        formula: Optional[FormulaSpec] = None,
+        formula: Optional[formulaic.FormulaSpec] = None,
         interaction_separator: str = ":",
         categorical_format: str = "{name}[{category}]",
         cat_missing_method: str = "fail",
@@ -835,7 +185,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             if self.fit_intercept:
                 coef = np.concatenate((np.array([intercept]), coef))
             if self._center_predictors:
-                _standardize_warm_start(coef, col_means, col_stds)  # type: ignore
+                standardize_warm_start(coef, col_means, col_stds)  # type: ignore
 
         elif self.start_params is None:
             if self.fit_intercept:
@@ -851,7 +201,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
                 )
 
         else:  # assign given array as start values
-            coef = check_array(
+            coef = skl.utils.check_array(
                 self.start_params,
                 accept_sparse=False,
                 ensure_2d=False,
@@ -868,7 +218,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
                 )
 
             if self._center_predictors:
-                _standardize_warm_start(coef, col_means, col_stds)  # type: ignore
+                standardize_warm_start(coef, col_means, col_stds)  # type: ignore
 
         # If starting values are outside the specified bounds (if set),
         # bring the starting value exactly at the bound.
@@ -891,7 +241,9 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         return coef
 
     def _convert_from_pandas(
-        self, df: pd.DataFrame, context: Optional[Mapping[str, Any]] = None
+        self,
+        df: pd.DataFrame,
+        context: Optional[Mapping[str, Any]] = None,
     ) -> tm.MatrixBase:
         """Convert a pandas data frame to a tabmat matrix."""
         if hasattr(self, "X_model_spec_"):
@@ -900,14 +252,14 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         cat_missing_method_after_alignment = getattr(self, "cat_missing_method", "fail")
 
         if hasattr(self, "feature_dtypes_"):
-            df = _align_df_categories(
+            df = align_df_categories(
                 df,
                 self.feature_dtypes_,
                 getattr(self, "has_missing_category_", {}),
                 cat_missing_method_after_alignment,
             )
             if cat_missing_method_after_alignment == "convert":
-                df = _add_missing_categories(
+                df = add_missing_categories(
                     df=df,
                     dtypes=self.feature_dtypes_,
                     feature_names=self.feature_names_,
@@ -1192,7 +544,8 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             P1 = P1_no_alpha * alpha
             P2 = P2_no_alpha * alpha
 
-            tic = perf_counter()
+            tic = time.perf_counter()
+
             coef = self._solve(
                 X=X,
                 y=y,
@@ -1206,7 +559,8 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
                 A_ineq=A_ineq,
                 b_ineq=b_ineq,
             )
-            toc = perf_counter()
+
+            toc = time.perf_counter()
 
             if self.verbose > 0:
                 print(
@@ -1218,7 +572,10 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         return self.coef_path_
 
     def report_diagnostics(
-        self, *, full_report: bool = False, custom_columns: Optional[Iterable] = None
+        self,
+        *,
+        full_report: bool = False,
+        custom_columns: Optional[Sequence] = None,
     ) -> None:
         """Print diagnostics to ``stdout``.
 
@@ -1246,7 +603,10 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             print(diagnostics)
 
     def get_formatted_diagnostics(
-        self, *, full_report: bool = False, custom_columns: Optional[Iterable] = None
+        self,
+        *,
+        full_report: bool = False,
+        custom_columns: Optional[Sequence] = None,
     ) -> Union[str, pd.DataFrame]:
         """Get formatted diagnostics which can be printed with report_diagnostics.
 
@@ -1347,7 +707,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         array, shape (n_samples, n_alphas)
             The linear predictor.
         """
-        check_is_fitted(self, "coef_")
+        skl.utils.validation.check_is_fitted(self, "coef_")
 
         if (alpha is not None) and (alpha_index is not None):
             raise ValueError("Please specify only one of {alpha_index, alpha}.")
@@ -1357,10 +717,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             alpha_index = [self._find_alpha_index(a) for a in alpha]  # type: ignore
 
         if isinstance(X, pd.DataFrame):
-            captured_context = capture_context(
-                context + 1 if isinstance(context, int) else context
-            )
-            X = self._convert_from_pandas(X, context=captured_context)
+            X = self._convert_from_pandas(X, context=capture_context(context))
 
         X = check_array_tabmat_compliant(
             X,
@@ -1449,20 +806,19 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             Predicted values times ``sample_weight``.
         """
         if isinstance(X, pd.DataFrame):
-            captured_context = capture_context(
-                context + 1 if isinstance(context, int) else context
-            )
-            X = self._convert_from_pandas(X, context=captured_context)
+            X = self._convert_from_pandas(X, context=capture_context(context))
 
         eta = self.linear_predictor(
-            X, offset=offset, alpha_index=alpha_index, alpha=alpha
+            X, offset=offset, alpha_index=alpha_index, alpha=alpha, context=context
         )
-        mu = get_link(self.link, get_family(self.family)).inverse(eta)
+
+        mu = self.link_instance.inverse(eta)
 
         if sample_weight is None:
             return mu
 
-        sample_weight = _check_weights(sample_weight, X.shape[0], X.dtype)
+        sample_weight = check_weights(sample_weight, X.shape[0], X.dtype)
+
         return mu * sample_weight
 
     def coef_table(
@@ -1533,9 +889,6 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             names = self.feature_names_
             beta = self.coef_
 
-        captured_context = capture_context(
-            context + 1 if isinstance(context, int) else context
-        )
         if (X is None) and (getattr(self, "covariance_matrix_", None) is None):
             return pd.Series(beta, index=names, name="coef")
 
@@ -1549,7 +902,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             robust=robust,
             clusters=clusters,
             expected_information=expected_information,
-            context=captured_context,
+            context=capture_context(context),
         )
 
         significance_level = 1 - confidence_level
@@ -1675,10 +1028,6 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
                 f"Received {num_lhs_specs} specifications."
             )
 
-        captured_context = capture_context(
-            context + 1 if isinstance(context, int) else context
-        )
-
         kwargs = {
             "X": X,
             "y": y,
@@ -1689,7 +1038,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             "robust": robust,
             "clusters": clusters,
             "expected_information": expected_information,
-            "context": captured_context,
+            "context": capture_context(context),
         }
 
         if R is not None:
@@ -1795,14 +1144,17 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         else:
             names = self.feature_names_
 
-        parser = LinearConstraintParser(names)
+        parser = formulaic.utils.constraints.LinearConstraintParser(names)
 
         R, r = parser.get_matrix(formula)
 
         return self._wald_test_matrix(R=R, r=r, **kwargs)
 
     def _wald_test_term_names(
-        self, terms: Union[str, list[str]], values: Optional[Sequence] = None, **kwargs
+        self,
+        terms: Union[str, list[str]],
+        values: Optional[Sequence] = None,
+        **kwargs,
     ) -> WaldTestResult:
         """
         Perform a Wald test for the hypothesis that the coefficients of the
@@ -1903,25 +1255,21 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             no context is added. Set ``context=0`` to make the calling scope
             available.
         """
-        captured_context = capture_context(
-            context + 1 if isinstance(context, int) else context
+        covariance_matrix = self.covariance_matrix(
+            X=X,
+            y=y,
+            sample_weight=sample_weight,
+            offset=offset,
+            mu=mu,
+            dispersion=dispersion,
+            robust=robust,
+            clusters=clusters,
+            expected_information=expected_information,
+            store_covariance_matrix=store_covariance_matrix,
+            context=capture_context(context),
         )
 
-        return np.sqrt(
-            self.covariance_matrix(
-                X=X,
-                y=y,
-                sample_weight=sample_weight,
-                offset=offset,
-                mu=mu,
-                dispersion=dispersion,
-                robust=robust,
-                clusters=clusters,
-                expected_information=expected_information,
-                store_covariance_matrix=store_covariance_matrix,
-                context=captured_context,
-            ).diagonal()
-        )
+        return np.sqrt(covariance_matrix.diagonal())
 
     def covariance_matrix(
         self,
@@ -2035,10 +1383,6 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         """
         self.covariance_matrix_: Union[np.ndarray, None]
 
-        captured_context = capture_context(
-            context + 1 if isinstance(context, int) else context
-        )
-
         if robust is None:
             _robust = getattr(self, "robust", True)
         else:
@@ -2107,7 +1451,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
                 # This has to go first because X is modified in the next line
 
             if isinstance(X, pd.DataFrame):
-                X = self._convert_from_pandas(X, context=captured_context)
+                X = self._convert_from_pandas(X, context=capture_context(context))
 
             X, y = check_X_y_tabmat_compliant(
                 X,
@@ -2125,13 +1469,13 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             if sparse.issparse(X) and not isinstance(X, tm.SparseMatrix):
                 X = tm.SparseMatrix(X)
 
-            sample_weight = _check_weights(
+            sample_weight = check_weights(
                 sample_weight,
                 y.shape[0],
                 X.dtype,
                 force_all_finite=self.force_all_finite,
             )
-            offset = _check_offset(offset, y.shape[0], X.dtype)
+            offset = check_offset(offset, y.shape[0], X.dtype)
 
         sum_weights = np.sum(sample_weight)  # type: ignore
 
@@ -2149,7 +1493,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
             )
 
         if (
-            np.linalg.cond(_safe_toarray(X.sandwich(np.ones(X.shape[0]))))
+            np.linalg.cond(safe_toarray(X.sandwich(np.ones(X.shape[0]))))
             > 1 / sys.float_info.epsilon**2
         ):
             raise np.linalg.LinAlgError(
@@ -2192,7 +1536,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
                 correction = sum_weights / (
                     sum_weights - self.n_features_in_ - int(self.fit_intercept)
                 )
-            vcov = linalg.solve(oim, linalg.solve(oim, _safe_toarray(inner_part)).T)
+            vcov = linalg.solve(oim, linalg.solve(oim, safe_toarray(inner_part)).T)
             vcov *= correction
         else:
             fisher = self._family_instance._fisher_information(
@@ -2204,7 +1548,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
                 dispersion,
                 self.fit_intercept,
             )
-            vcov = linalg.inv(_safe_toarray(fisher))
+            vcov = linalg.inv(safe_toarray(fisher))
             vcov *= sum_weights / (
                 sum_weights - self.n_features_in_ - int(self.fit_intercept)
             )
@@ -2265,20 +1609,15 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         float
             D^2 of self.predict(X) w.r.t. y.
         """
-        # Note, default score defined in RegressorMixin is R^2 score.
-        # TODO: make D^2 a score function in module metrics (and thereby get
-        #       input validation and so on)
-        captured_context = capture_context(
-            context + 1 if isinstance(context, int) else context
-        )
+        sample_weight = check_weights(sample_weight, y.shape[0], y.dtype)
 
-        sample_weight = _check_weights(sample_weight, y.shape[0], y.dtype)
-        mu = self.predict(X, offset=offset, context=captured_context)
-        family = get_family(self.family)
-        dev = family.deviance(y, mu, sample_weight=sample_weight)
+        mu = self.predict(X, offset=offset, context=context)
         y_mean = np.average(y, weights=sample_weight)
-        dev_null = family.deviance(y, y_mean, sample_weight=sample_weight)
-        return 1.0 - dev / dev_null
+
+        dev = self.family_instance.deviance(y, mu, sample_weight=sample_weight)
+        dev_null = self.family_instance.deviance(y, y_mean, sample_weight=sample_weight)
+
+        return 1 - dev / dev_null
 
     def _validate_hyperparameters(self) -> None:
         if not isinstance(self.fit_intercept, bool):
@@ -2408,7 +1747,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
 
         if isinstance(X, pd.DataFrame):
             if hasattr(self, "formula") and self.formula is not None:
-                lhs, rhs = _parse_formula(
+                lhs, rhs = parse_formula(
                     self.formula, include_intercept=self.fit_intercept
                 )
 
@@ -2471,10 +1810,10 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
                 }
 
                 if any(X.dtypes == "category"):
-                    P1 = _expand_categorical_penalties(
+                    P1 = expand_categorical_penalties(
                         self.P1, X, drop_first, self.has_missing_category_
                     )
-                    P2 = _expand_categorical_penalties(
+                    P2 = expand_categorical_penalties(
                         self.P2, X, drop_first, self.has_missing_category_
                     )
 
@@ -2494,7 +1833,7 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
                 "requires y to be passed, but the target y is None."
             )
 
-        if not _is_contiguous(X):
+        if not is_contiguous(X):
             if self.copy_X is not None and not self.copy_X:
                 raise ValueError(
                     "The X matrix is noncontiguous and copy_X = False."
@@ -2550,14 +1889,14 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         # mixed-precision numbers
         y = np.asarray(y, dtype=X.dtype)
 
-        sample_weight = _check_weights(
+        sample_weight = check_weights(
             sample_weight,
             y.shape[0],  # type: ignore
             X.dtype,
             force_all_finite=force_all_finite,
         )
 
-        offset = _check_offset(offset, y.shape[0], X.dtype)  # type: ignore
+        offset = check_offset(offset, y.shape[0], X.dtype)  # type: ignore
 
         # IMPORTANT NOTE: Since we want to minimize
         # 1/(2*sum(sample_weight)) * deviance + L1 + L2,
@@ -2576,976 +1915,203 @@ class GeneralizedLinearRegressorBase(RegressorMixin, BaseEstimator):
         return X, y, sample_weight, offset, weights_sum, P1, P2  # type: ignore
 
 
-class GeneralizedLinearRegressor(GeneralizedLinearRegressorBase):
-    """Regression via a Generalized Linear Model (GLM) with penalties.
-
-    GLMs based on a reproductive Exponential Dispersion Model (EDM) aimed at
-    fitting and predicting the mean of the target ``y`` as ``mu=h(X*w)``.
-    Therefore, the fit minimizes the following objective function with combined
-    L1 and L2 priors as regularizer::
-
-            1/(2*sum(s)) * deviance(y, h(X*w); s)
-            + alpha * l1_ratio * ||P1*w||_1
-            + 1/2 * alpha * (1 - l1_ratio) * w*P2*w
-
-    with inverse link function ``h`` and ``s=sample_weight``.
-    Note that, for ``alpha=0`` the unregularized GLM is recovered.
-    This is not the default behavior (see ``alpha`` parameter description for details).
-    Additionally, for ``sample_weight=None``, one has ``s_i=1`` and
-    ``sum(s)=n_samples``. For ``P1=P2='identity'``, the penalty is the elastic net::
-
-            alpha * l1_ratio * ||w||_1 + 1/2 * alpha * (1 - l1_ratio) * ||w||_2^2.
-
-    If you are interested in controlling the L1 and L2 penalties separately,
-    keep in mind that this is equivalent to::
-
-            a * L1 + b * L2,
-
-    where::
-
-            alpha = a + b and l1_ratio = a / (a + b).
-
-    The parameter ``l1_ratio`` corresponds to alpha in the R package glmnet,
-    while ``alpha`` corresponds to the lambda parameter in glmnet.
-    Specifically, ``l1_ratio = 1`` is the lasso penalty.
-
-    Read more in :doc:`background<background>`.
-
-    Parameters
-    ----------
-    alpha : {float, array-like}, optional (default=None)
-        Constant that multiplies the penalty terms and thus determines the
-        regularization strength. If ``alpha_search`` is ``False`` (the default),
-        then ``alpha`` must be a scalar or None (equivalent to ``alpha=0``).
-        If ``alpha_search`` is ``True``, then ``alpha`` must be an iterable or
-        ``None``. See ``alpha_search`` to find how the regularization path is
-        set if ``alpha`` is ``None``. See the notes for the exact mathematical
-        meaning of this parameter. ``alpha=0`` is equivalent to unpenalized
-        GLMs. In this case, the design matrix ``X`` must have full column rank
-        (no collinearities).
-
-    l1_ratio : float, optional (default=0)
-        The elastic net mixing parameter, with ``0 <= l1_ratio <= 1``. For
-        ``l1_ratio = 0``, the penalty is an L2 penalty. ``For l1_ratio = 1``, it
-        is an L1 penalty.  For ``0 < l1_ratio < 1``, the penalty is a
-        combination of L1 and L2.
-
-    P1 : {'identity', array-like, None}, shape (n_features,), optional
-         (default='identity')
-        This array controls the strength of the regularization for each coefficient
-        independently. A high value will lead to higher regularization while a value of
-        zero will remove the regularization on this parameter.
-        Note that ``n_features = X.shape[1]``. If ``X`` is a pandas DataFrame
-        with a categorical dtype and P1 has the same size as the number of columns,
-        the penalty of the categorical column will be applied to all the levels of
-        the categorical.
-
-    P2 : {'identity', array-like, sparse matrix, None}, shape (n_features,) \
-            or (n_features, n_features), optional (default='identity')
-        With this option, you can set the P2 matrix in the L2 penalty
-        ``w*P2*w``. This gives a fine control over this penalty (Tikhonov
-        regularization). A 2d array is directly used as the square matrix P2. A
-        1d array is interpreted as diagonal (square) matrix. The default
-        ``'identity'`` and ``None`` set the identity matrix, which gives the usual
-        squared L2-norm. If you just want to exclude certain coefficients, pass a 1d
-        array filled with 1 and 0 for the coefficients to be excluded. Note that P2 must
-        be positive semi-definite. If ``X`` is a pandas DataFrame with a categorical
-        dtype and P2 has the same size as the number of columns, the penalty of the
-        categorical column will be applied to all the levels of the categorical. Note
-        that if P2 is two-dimensional, its size needs to be of the same length as the
-        expanded ``X`` matrix.
-
-    fit_intercept : bool, optional (default=True)
-        Specifies if a constant (a.k.a. bias or intercept) should be
-        added to the linear predictor (``X * coef + intercept``).
-
-    family : str or ExponentialDispersionModel, optional (default='normal')
-        The distributional assumption of the GLM, i.e. the loss function to
-        minimize. If a string, one of: ``'binomial'``, ``'gamma'``,
-        ``'gaussian'``, ``'inverse.gaussian'``, ``'normal'``, ``'poisson'``,
-        ``'tweedie'`` or ``'negative.binomial'``. Note that ``'tweedie'`` sets
-        the power of the Tweedie distribution to 1.5; to use another value,
-        specify it in parentheses (e.g., ``'tweedie (1.5)'``). The same applies
-        for ``'negative.binomial'`` and theta parameter.
-
-    link : {'auto', 'identity', 'log', 'logit', 'cloglog'} oe Link, \
-            optional (default='auto')
-        The link function of the GLM, i.e. mapping from linear
-        predictor (``X * coef``) to expectation (``mu``). Option ``'auto'`` sets
-        the link depending on the chosen family as follows:
-
-        - ``'identity'`` for family ``'normal'``
-        - ``'log'`` for families ``'poisson'``, ``'gamma'``,
-          ``'inverse.gaussian'`` and ``'negative.binomial'``.
-        - ``'logit'`` for family ``'binomial'``
-
-    solver : {'auto', 'irls-cd', 'irls-ls', 'lbfgs', 'trust-constr'}, \
-            optional (default='auto')
-        Algorithm to use in the optimization problem:
-
-        - ``'auto'``: ``'irls-ls'`` if ``l1_ratio`` is zero and ``'irls-cd'`` otherwise.
-        - ``'irls-cd'``: Iteratively reweighted least squares with a coordinate
-          descent inner solver. This can deal with L1 as well as L2 penalties.
-          Note that in order to avoid unnecessary memory duplication of X in the
-          ``fit`` method, ``X`` should be directly passed as a
-          Fortran-contiguous Numpy array or sparse CSC matrix.
-        - ``'irls-ls'``: Iteratively reweighted least squares with a least
-          squares inner solver. This algorithm cannot deal with L1 penalties.
-        - ``'lbfgs'``: Scipy's L-BFGS-B optimizer. It cannot deal with L1
-          penalties.
-        - ``'trust-constr'``: Calls
-          ``scipy.optimize.minimize(method='trust-constr')``. It cannot deal
-          with L1 penalties. This solver can optimize problems with inequality
-          constraints, passed via ``A_ineq`` and ``b_ineq``. It will be selected
-          automatically when inequality constraints are set and
-          ``solver='auto'``. Note that using this method can lead to
-          significantly increased runtimes by a factor of ten or higher.
-
-    max_iter : int, optional (default=100)
-        The maximal number of iterations for solver algorithms.
-
-    max_inner_iter: int, optional (default=100000)
-        The maximal number of iterations for the inner solver in the IRLS-CD
-        algorithm. This parameter is only used when ``solver='irls-cd'``.
-
-    gradient_tol : float, optional (default=None)
-        Stopping criterion. If ``None``, solver-specific defaults will be used.
-        The default value for most solvers is ``1e-4``, except for
-        ``'trust-constr'``, which requires more conservative convergence
-        settings and has a default value of ``1e-8``.
-
-        For the IRLS-LS, L-BFGS and trust-constr solvers, the iteration
-        will stop when ``max{|g_i|, i = 1, ..., n} <= tol``, where ``g_i`` is
-        the ``i``-th component of the gradient (derivative) of the objective
-        function. For the CD solver, convergence is reached when
-        ``sum_i(|minimum norm of g_i|)``, where ``g_i`` is the subgradient of
-        the objective and the minimum norm of ``g_i`` is the element of the
-        subgradient with the smallest L2 norm.
-
-        If you wish to only use a step-size tolerance, set ``gradient_tol``
-        to a very small number.
-
-    step_size_tol: float, optional (default=None)
-        Alternative stopping criterion. For the IRLS-LS and IRLS-CD solvers, the
-        iteration will stop when the L2 norm of the step size is less than
-        ``step_size_tol``. This stopping criterion is disabled when
-        ``step_size_tol`` is ``None``.
-
-    hessian_approx: float, optional (default=0.0)
-        The threshold below which data matrix rows will be ignored for updating
-        the Hessian. See the algorithm documentation for the IRLS algorithm
-        for further details.
-
-    warm_start : bool, optional (default=False)
-        Whether to reuse the solution of the previous call to ``fit``
-        as initialization for ``coef_`` and ``intercept_`` (supersedes
-        ``start_params``). If ``False`` or if the attribute ``coef_`` does not
-        exist (first call to ``fit``), ``start_params`` sets the start values
-        for ``coef_`` and ``intercept_``.
-
-    alpha_search : bool, optional (default=False)
-        Whether to search along the regularization path for the best alpha.
-        When set to ``True``, ``alpha`` should either be ``None`` or an
-        iterable. To determine the regularization path, the following sequence
-        is used:
-
-        1. If ``alpha`` is an iterable, use it directly. All other parameters
-           governing the regularization path are ignored.
-        2. If ``min_alpha`` is set, create a path from ``min_alpha`` to the
-           lowest alpha such that all coefficients are zero.
-        3. If ``min_alpha_ratio`` is set, create a path where the ratio of
-           ``min_alpha / max_alpha = min_alpha_ratio``.
-        4. If none of the above parameters are set, use a ``min_alpha_ratio`` of
-           ``1e-6``.
-
-    alphas : DEPRECATED. Use ``alpha`` instead.
-
-    n_alphas : int, optional (default=100)
-        Number of alphas along the regularization path
-
-    min_alpha_ratio : float, optional (default=None)
-        Length of the path. ``min_alpha_ratio=1e-6`` means that
-        ``min_alpha / max_alpha = 1e-6``. If ``None``, ``1e-6`` is used.
-
-    min_alpha : float, optional (default=None)
-        Minimum alpha to estimate the model with. The grid will then be created
-        over ``[max_alpha, min_alpha]``.
-
-    start_params : array-like, shape (n_features*,), optional (default=None)
-        Relevant only if ``warm_start`` is ``False`` or if ``fit`` is called
-        for the first time (so that ``self.coef_`` does not exist yet). If
-        ``None``, all coefficients are set to zero and the start value for the
-        intercept is the weighted average of ``y`` (If ``fit_intercept`` is
-        ``True``). If an array, used directly as start values; if
-        ``fit_intercept`` is ``True``, its first element is assumed to be the
-        start value for the ``intercept_``. Note that
-        ``n_features* = X.shape[1] + fit_intercept``, i.e. it includes the
-        intercept.
-
-    selection : str, optional (default='cyclic')
-        For the CD solver 'cd', the coordinates (features) can be updated in
-        either cyclic or random order. If set to ``'random'``, a random
-        coefficient is updated every iteration rather than looping over features
-        sequentially in the same order, which often leads to significantly
-        faster convergence, especially when ``gradient_tol`` is higher than
-        ``1e-4``.
-
-    random_state : int or RandomState, optional (default=None)
-        The seed of the pseudo random number generator that selects a random
-        feature to be updated for the CD solver. If an integer, ``random_state``
-        is the seed used by the random number generator; if a
-        :class:`RandomState` instance, ``random_state`` is the random number
-        generator; if ``None``, the random number generator is the
-        :class:`RandomState` instance used by ``np.random``. Used when
-        ``selection`` is ``'random'``.
-
-    copy_X : bool, optional (default=None)
-        Whether to copy ``X``. Since ``X`` is never modified by
-        :class:`GeneralizedLinearRegressor`, this is unlikely to be needed; this
-        option exists mainly for compatibility with other scikit-learn
-        estimators. If ``False``, ``X`` will not be copied and there will be an
-        error if you pass an ``X`` in the wrong format, such as providing
-        integer ``X`` and float ``y``. If ``None``, ``X`` will not be copied
-        unless it is in the wrong format.
-
-    check_input : bool, optional (default=True)
-        Whether to bypass several checks on input: ``y`` values in range of
-        ``family``, ``sample_weight`` non-negative, ``P2`` positive
-        semi-definite. Don't use this parameter unless you know what you are
-        doing.
-
-    verbose : int, optional (default=0)
-        For the IRLS solver, any positive number will result in a pretty
-        progress bar showing convergence. This features requires having the
-        tqdm package installed. For the L-BFGS and ``'trust-constr'`` solvers,
-        set ``verbose`` to any positive number for verbosity.
-
-    scale_predictors: bool, optional (default=False)
-        If ``True``, scale all predictors to have standard deviation one.
-        Should be set to ``True`` if ``alpha > 0`` and if you want coefficients
-        to be penalized equally.
-
-        Reported coefficient estimates are always at the original scale.
-
-        Advanced developer note: Internally, predictors are always rescaled for
-        computational reasons, but this only affects results if
-        ``scale_predictors`` is ``True``.
-
-    lower_bounds : array-like, shape (n_features,), optional (default=None)
-        Set a lower bound for the coefficients. Setting bounds forces the use
-        of the coordinate descent solver (``'irls-cd'``).
-
-    upper_bounds : array-like, shape=(n_features,), optional (default=None)
-        See ``lower_bounds``.
-
-    A_ineq : array-like, shape=(n_constraints, n_features), optional (default=None)
-        Constraint matrix for linear inequality constraints of the form
-        ``A_ineq w <= b_ineq``. Setting inequality constraints forces the use
-        of the local gradient-based solver ``'trust-constr'``, which may
-        increase runtime significantly. Note that the constraints only apply
-        to coefficients related to features in ``X``. If you want to constrain
-        the intercept, add it to the feature matrix ``X`` manually and set
-        ``fit_intercept==False``.
-
-    b_ineq : array-like, shape=(n_constraints,), optional (default=None)
-        Constraint vector for linear inequality constraints of the form
-        ``A_ineq w <= b_ineq``. Refer to the documentation of ``A_ineq`` for
-        details.
-
-    drop_first : bool, optional (default = False)
-        If ``True``, drop the first column when encoding categorical variables.
-        Set this to True when ``alpha=0`` and ``solver='auto'`` to prevent an error
-        due to a singular feature matrix. In the case of using a formula with
-        interactions, setting this argument to ``True`` ensures structural
-        full-rankness (it is equivalent to ``ensure_full_rank`` in formulaic and
-        tabmat).
-
-    robust : bool, optional (default = False)
-        If true, then robust standard errors are computed by default.
-
-    expected_information : bool, optional (default = False)
-        If true, then the expected information matrix is computed by default.
-        Only relevant when computing robust standard errors.
-
-    formula : formulaic.FormulaSpec
-        A formula accepted by formulaic. It can either be a one-sided formula, in
-        which case ``y`` must be specified in ``fit``, or a two-sided formula, in
-        which case ``y`` must be ``None``.
-
-    interaction_separator: str, default=":"
-        The separator between the names of interacted variables.
-
-    categorical_format : str, optional, default='{name}[{category}]'
-        Format string for categorical features. The format string should
-        contain the placeholder ``{name}`` for the feature name and
-        ``{category}`` for the category name. Only used if ``X`` is a pandas
-        DataFrame.
-
-    cat_missing_method: str {'fail'|'zero'|'convert'}, default='fail'
-        How to handle missing values in categorical columns. Only used if ``X``
-        is a pandas data frame.
-        - if 'fail', raise an error if there are missing values
-        - if 'zero', missing values will represent all-zero indicator columns.
-        - if 'convert', missing values will be converted to the ``cat_missing_name``
-          category.
-
-    cat_missing_name: str, default='(MISSING)'
-        Name of the category to which missing values will be converted if
-        ``cat_missing_method='convert'``.  Only used if ``X`` is a pandas data frame.
-
-    Attributes
-    ----------
-    coef_ : numpy.array, shape (n_features,)
-        Estimated coefficients for the linear predictor (X*coef_+intercept_) in
-        the GLM.
-
-    intercept_ : float
-        Intercept (a.k.a. bias) added to linear predictor.
-
-    n_iter_ : int
-        Actual number of iterations used in solver.
-
-    col_means_: array, shape (n_features,)
-        The means of the columns of the design matrix ``X``.
-
-    col_stds_: array, shape (n_features,)
-        The standard deviations of the columns of the design matrix ``X``.
-
-    Notes
-    -----
-    The fit itself does not need outcomes to be from an EDM, but only assumes
-    the first two moments to be
-    :math:`\\mu_i \\equiv \\mathrm{E}(y_i) = h(x_i' w)` and
-    :math:`\\mathrm{var}(y_i) = (\\phi / s_i) v(\\mu_i)`. The unit
-    variance function :math:`v(\\mu_i)` is a property of and given by the
-    specific EDM; see :doc:`background<background>`.
-
-    The parameters :math:`w` (``coef_`` and ``intercept_``) are estimated by
-    minimizing the deviance plus penalty term, which is equivalent to
-    (penalized) maximum likelihood estimation.
-
-    If the target ``y`` is a ratio, appropriate sample weights ``s`` should be
-    provided. As an example, consider Poisson distributed counts ``z``
-    (integers) and weights ``s = exposure`` (time, money, persons years, ...).
-    Then you fit ``y ≡ z/s``, i.e.
-    ``GeneralizedLinearModel(family='poisson').fit(X, y, sample_weight=s)``. The
-    weights are necessary for the right (finite sample) mean. Consider
-    :math:`\\bar{y} = \\sum_i s_i y_i / \\sum_i s_i`: in this case, one might
-    say that :math:`y` follows a 'scaled' Poisson distribution. The same holds
-    for other distributions.
-
-    References
-    ----------
-    For the coordinate descent implementation:
-        * Guo-Xun Yuan, Chia-Hua Ho, Chih-Jen Lin
-          An Improved GLMNET for L1-regularized Logistic Regression,
-          Journal of Machine Learning Research 13 (2012) 1999-2030
-          https://www.csie.ntu.edu.tw/~cjlin/papers/l1_glmnet/long-glmnet.pdf
+def get_family(
+    family: Union[str, ExponentialDispersionModel],
+) -> ExponentialDispersionModel:
+    if isinstance(family, ExponentialDispersionModel):
+        return family
+
+    name_to_dist = {
+        "binomial": BinomialDistribution(),
+        "gamma": GammaDistribution(),
+        "gaussian": NormalDistribution(),
+        "inverse.gaussian": InverseGaussianDistribution(),
+        "normal": NormalDistribution(),
+        "poisson": PoissonDistribution(),
+        "tweedie": TweedieDistribution(1.5),
+        "negative.binomial": NegativeBinomialDistribution(1.0),
+    }
+
+    if family in name_to_dist:
+        return name_to_dist[family]
+
+    custom_tweedie = re.search(r"tweedie\s?\((.+)\)", family)
+
+    if custom_tweedie:
+        return TweedieDistribution(float(custom_tweedie.group(1)))
+
+    custom_negative_binomial = re.search(r"negative.binomial\s?\((.+)\)", family)
+
+    if custom_negative_binomial:
+        return NegativeBinomialDistribution(float(custom_negative_binomial.group(1)))
+
+    raise ValueError(
+        "The family must be an instance of class ExponentialDispersionModel or an "
+        f"element of {sorted(name_to_dist.keys())}; got (family={family})."
+    )
+
+
+def get_link(link: Union[str, Link], family: ExponentialDispersionModel) -> Link:
     """
+    For the Tweedie distribution, this code follows actuarial best practices
+    regarding link functions. Note that these links are sometimes not canonical:
+        - identity for normal (``p = 0``);
+        - no convention for ``p < 0``, so let's leave it as identity;
+        - log otherwise.
+    """
+    if isinstance(link, Link):
+        return link
 
-    def __init__(
-        self,
-        *,
-        alpha=None,
-        l1_ratio=0,
-        P1: Optional[Union[str, np.ndarray]] = "identity",
-        P2: Optional[Union[str, np.ndarray, sparse.spmatrix]] = "identity",
-        fit_intercept=True,
-        family: Union[str, ExponentialDispersionModel] = "normal",
-        link: Union[str, Link] = "auto",
-        solver: str = "auto",
-        max_iter=100,
-        max_inner_iter=100000,
-        gradient_tol: Optional[float] = None,
-        step_size_tol: Optional[float] = None,
-        hessian_approx: float = 0.0,
-        warm_start: bool = False,
-        alpha_search: bool = False,
-        alphas: Optional[np.ndarray] = None,
-        n_alphas: int = 100,
-        min_alpha_ratio: Optional[float] = None,
-        min_alpha: Optional[float] = None,
-        start_params: Optional[np.ndarray] = None,
-        selection: str = "cyclic",
-        random_state=None,
-        copy_X: Optional[bool] = None,
-        check_input: bool = True,
-        verbose=0,
-        scale_predictors: bool = False,
-        lower_bounds: Optional[np.ndarray] = None,
-        upper_bounds: Optional[np.ndarray] = None,
-        A_ineq: Optional[np.ndarray] = None,
-        b_ineq: Optional[np.ndarray] = None,
-        force_all_finite: bool = True,
-        drop_first: bool = False,
-        robust: bool = True,
-        expected_information: bool = False,
-        formula: Optional[FormulaSpec] = None,
-        interaction_separator: str = ":",
-        categorical_format: str = "{name}[{category}]",
-        cat_missing_method: str = "fail",
-        cat_missing_name: str = "(MISSING)",
-    ):
-        self.alphas = alphas
-        self.alpha = alpha
-        super().__init__(
-            l1_ratio=l1_ratio,
-            P1=P1,
-            P2=P2,
-            fit_intercept=fit_intercept,
-            family=family,
-            link=link,
-            solver=solver,
-            max_iter=max_iter,
-            max_inner_iter=max_inner_iter,
-            gradient_tol=gradient_tol,
-            step_size_tol=step_size_tol,
-            hessian_approx=hessian_approx,
-            warm_start=warm_start,
-            alpha_search=alpha_search,
-            n_alphas=n_alphas,
-            min_alpha=min_alpha,
-            min_alpha_ratio=min_alpha_ratio,
-            start_params=start_params,
-            selection=selection,
-            random_state=random_state,
-            copy_X=copy_X,
-            check_input=check_input,
-            verbose=verbose,
-            scale_predictors=scale_predictors,
-            lower_bounds=lower_bounds,
-            upper_bounds=upper_bounds,
-            A_ineq=A_ineq,
-            b_ineq=b_ineq,
-            force_all_finite=force_all_finite,
-            drop_first=drop_first,
-            robust=robust,
-            expected_information=expected_information,
-            formula=formula,
-            interaction_separator=interaction_separator,
-            categorical_format=categorical_format,
-            cat_missing_method=cat_missing_method,
-            cat_missing_name=cat_missing_name,
+    if (link is None) or (link == "auto"):
+        if tweedie_representation := family.to_tweedie(safe=False):
+            if tweedie_representation.power <= 0:
+                return IdentityLink()
+            return LogLink()
+        if isinstance(family, GeneralizedHyperbolicSecant):
+            return IdentityLink()
+        if isinstance(family, BinomialDistribution):
+            return LogitLink()
+        if isinstance(family, NegativeBinomialDistribution):
+            return LogLink()
+        raise ValueError(
+            "No default link known for the specified distribution family. "
+            "Please set link manually, i.e. not to 'auto'. "
+            f"Got (link='auto', family={family.__class__.__name__})."
         )
 
-    def _validate_hyperparameters(self) -> None:
-        if self.alpha_search:
-            if not isinstance(self.alpha, Iterable) and self.alpha is not None:
-                raise ValueError(
-                    "`alpha` should be an Iterable or None when `alpha_search`"
-                    " is True"
-                )
-            if self.alpha is not None and (
-                (np.asarray(self.alpha) < 0).any()
-                or not np.issubdtype(np.asarray(self.alpha).dtype, np.number)
-            ):
-                raise ValueError("`alpha` must contain only non-negative numbers")
-        if not self.alpha_search:
-            if not np.isscalar(self.alpha) and self.alpha is not None:
-                raise ValueError(
-                    "`alpha` should be a scalar or None when `alpha_search`" " is False"
-                )
-            if self.alpha is not None and (
-                not isinstance(self.alpha, (int, float)) or self.alpha < 0
-            ):
-                raise ValueError(
-                    "Penalty term must be a non-negative number;"
-                    f" got (alpha={self.alpha})"  # type: ignore
-                )
+    mapping = {
+        "cloglog": CloglogLink(),
+        "identity": IdentityLink(),
+        "log": LogLink(),
+        "logit": LogitLink(),
+        "tweedie": TweedieLink(1.5),
+    }
 
-        if (
-            not np.isscalar(self.l1_ratio)
-            # check for numeric, i.e. not a string
-            or not np.issubdtype(np.asarray(self.l1_ratio).dtype, np.number)
-            or self.l1_ratio < 0  # type: ignore
-            or self.l1_ratio > 1  # type: ignore
-        ):
+    if link in mapping:
+        return mapping[link]
+    if custom_tweedie := re.search(r"tweedie\s?\((.+)\)", link):
+        return TweedieLink(float(custom_tweedie.group(1)))
+
+    raise ValueError(
+        "The link must be an instance of class Link or an element of "
+        "['auto', 'identity', 'log', 'logit', 'cloglog', 'tweedie']; "
+        f"got (link={link})."
+    )
+
+
+def setup_p1(
+    P1: Optional[Union[str, np.ndarray]],
+    X: Union[tm.MatrixBase, tm.StandardizedMatrix],
+    dtype,
+    alpha: float,
+    l1_ratio: float,
+) -> np.ndarray:
+    if not isinstance(X, (tm.MatrixBase, tm.StandardizedMatrix)):
+        raise TypeError
+
+    n_features = X.shape[1]
+
+    if isinstance(P1, str):
+        if P1 != "identity":
+            raise ValueError(f"P1 must be either 'identity' or an array; got {P1}.")
+        P1 = np.ones(n_features, dtype=dtype)
+    elif P1 is None:
+        P1 = np.ones(n_features, dtype=dtype)
+    else:
+        P1 = np.atleast_1d(P1)
+        try:
+            P1 = P1.astype(dtype, casting="safe", copy=False)  # type: ignore
+        except TypeError as e:
+            raise TypeError(
+                "The given P1 cannot be converted to a numeric array; "
+                f"got (P1.dtype={P1.dtype})."  # type: ignore
+            ) from e
+        if (P1.ndim != 1) or (P1.shape[0] != n_features):  # type: ignore
             raise ValueError(
-                "l1_ratio must be a number in interval [0, 1];"
-                f" got (l1_ratio={self.l1_ratio})"
-            )
-        super()._validate_hyperparameters()
-
-    def fit(
-        self,
-        X: ArrayLike,
-        y: Optional[ArrayLike] = None,
-        sample_weight: Optional[ArrayLike] = None,
-        offset: Optional[ArrayLike] = None,
-        *,
-        store_covariance_matrix: bool = False,
-        clusters: Optional[np.ndarray] = None,
-        # TODO: take out weights_sum (or use it properly)
-        weights_sum: Optional[float] = None,
-        context: Optional[Union[int, Mapping[str, Any]]] = None,
-    ):
-        """Fit a Generalized Linear Model.
-
-        Parameters
-        ----------
-        X : {array-like, sparse matrix}, shape (n_samples, n_features)
-            Training data. Note that a ``float32`` matrix is acceptable and will
-            result in the entire algorithm being run in 32-bit precision.
-            However, for problems that are poorly conditioned, this might result
-            in poor convergence or flawed parameter estimates. If a Pandas data
-            frame is provided, it may contain categorical columns. In that case,
-            a separate coefficient will be estimated for each category. No
-            category is omitted. This means that some regularization is required
-            to fit models with an intercept or models with several categorical
-            columns.
-
-        y : array-like, shape (n_samples,)
-            Target values.
-
-        sample_weight : array-like, shape (n_samples,), optional (default=None)
-            Individual weights w_i for each sample. Note that, for an
-            Exponential Dispersion Model (EDM), one has
-            :math:`\\mathrm{var}(y_i) = \\phi \\times v(mu) / w_i`.
-            If :math:`y_i \\sim EDM(\\mu, \\phi / w_i)`, then
-            :math:`\\sum w_i y_i / \\sum w_i \\sim EDM(\\mu, \\phi / \\sum w_i)`,
-            i.e. the mean of :math:`y` is a weighted average with weights equal
-            to ``sample_weight``.
-
-        offset: array-like, shape (n_samples,), optional (default=None)
-            Added to linear predictor. An offset of 3 will increase expected
-            ``y`` by 3 if the link is linear and will multiply expected ``y`` by
-            3 if the link is logarithmic.
-
-        store_covariance_matrix : bool, optional (default=False)
-            Whether to estimate and store the covariance matrix of the parameter
-            estimates. If ``True``, the covariance matrix will be available in the
-            ``covariance_matrix_`` attribute after fitting.
-
-        clusters : array-like, optional, default=None
-            Array with cluster membership. Clustered standard errors are
-            computed if clusters is not None.
-
-        context : Optional[Union[int, Mapping[str, Any]]], default=None
-            The context to add to the evaluation context of the formula with,
-            e.g., custom transforms. If an integer, the context is taken from
-            the stack frame of the caller at the given depth. Otherwise, a
-            mapping from variable names to values is expected. By default,
-            no context is added. Set ``context=0`` to make the calling scope
-            available.
-
-        weights_sum: float, optional (default=None)
-
-        Returns
-        -------
-        self
-        """
-
-        self._validate_hyperparameters()
-
-        captured_context = capture_context(
-            context + 1 if isinstance(context, int) else context
-        )
-
-        # NOTE: This function checks if all the entries in X and y are
-        # finite. That can be expensive. But probably worthwhile.
-        (
-            X,
-            y,
-            sample_weight,
-            offset,
-            weights_sum,
-            P1,
-            P2,
-        ) = self._set_up_and_check_fit_args(
-            X,
-            y,
-            sample_weight,
-            offset,
-            force_all_finite=self.force_all_finite,
-            context=captured_context,
-        )
-        assert isinstance(X, tm.MatrixBase)
-        assert isinstance(y, np.ndarray)
-
-        self._set_up_for_fit(y)
-
-        # 1.3 arguments to take special care ##################################
-        # P1, P2, start_params
-        stype = ["csc"] if self._solver == "irls-cd" else ["csc", "csr"]
-        P1_no_alpha = setup_p1(P1, X, X.dtype, 1, self.l1_ratio)
-        P2_no_alpha = setup_p2(P2, X, stype, X.dtype, 1, self.l1_ratio)
-
-        lower_bounds = check_bounds(self.lower_bounds, X.shape[1], X.dtype)
-        upper_bounds = check_bounds(self.upper_bounds, X.shape[1], X.dtype)
-
-        A_ineq, b_ineq = check_inequality_constraints(
-            self.A_ineq, self.b_ineq, n_features=X.shape[1], dtype=X.dtype
-        )
-
-        if (lower_bounds is not None) and (upper_bounds is not None):
-            if np.any(lower_bounds > upper_bounds):
-                raise ValueError("Upper bounds must be higher than lower bounds.")
-
-        # 1.4 additional validations ##########################################
-        if self.check_input:
-            # check if P2 is positive semidefinite
-            if not isinstance(self.P2, str):  # self.P2 != 'identity'
-                if not is_pos_semidef(P2_no_alpha):
-                    if P2_no_alpha.ndim == 1 or P2_no_alpha.shape[0] == 1:
-                        error = "1d array P2 must not have negative values."
-                    else:
-                        error = "P2 must be positive semi-definite."
-                    raise ValueError(error)
-            # TODO: if alpha=0 check that X is not rank deficient
-            # TODO: what else to check?
-
-        #######################################################################
-        # 2c. potentially rescale predictors
-        #######################################################################
-
-        (
-            X,
-            self.col_means_,
-            self.col_stds_,
-            lower_bounds,
-            upper_bounds,
-            A_ineq,
-            P1_no_alpha,
-            P2_no_alpha,
-        ) = _standardize(
-            X,
-            sample_weight,
-            self._center_predictors,
-            self.scale_predictors,
-            lower_bounds,
-            upper_bounds,
-            A_ineq,
-            P1_no_alpha,
-            P2_no_alpha,
-        )
-
-        #######################################################################
-        # 3. initialization of coef = (intercept_, coef_)                     #
-        #######################################################################
-
-        coef = self._get_start_coef(
-            X,
-            y,
-            sample_weight,
-            offset,
-            self.col_means_,
-            self.col_stds_,
-            dtype=[np.float64, np.float32],
-        )
-
-        #######################################################################
-        # 4. fit                                                              #
-        #######################################################################
-        if self.alpha_search:
-            if self.alphas is not None:
-                warnings.warn(
-                    "alphas is deprecated. Use alpha instead.", DeprecationWarning
-                )
-                self._alphas = self.alphas
-            elif self.alpha is None:
-                self._alphas = self._get_alpha_path(
-                    P1_no_alpha=P1_no_alpha, X=X, y=y, w=sample_weight, offset=offset
-                )
-            else:
-                self._alphas = self.alpha
-                if self.min_alpha is not None or self.min_alpha_ratio is not None:
-                    warnings.warn(
-                        "`alpha` is set. Ignoring `min_alpha` and `min_alpha_ratio`."
-                    )
-
-            coef = self._solve_regularization_path(
-                X=X,
-                y=y,
-                sample_weight=sample_weight,
-                P2_no_alpha=P2_no_alpha,
-                P1_no_alpha=P1_no_alpha,
-                alphas=self._alphas,
-                coef=coef,
-                offset=offset,
-                lower_bounds=lower_bounds,
-                upper_bounds=upper_bounds,
-                A_ineq=A_ineq,
-                b_ineq=b_ineq,
+                "P1 must be either 'identity' or a 1d array with the length of "
+                "X.shape[1] (either before or after categorical expansion); "
+                f"got (P1.shape[0]={P1.shape[0]})."  # type: ignore
             )
 
-            # intercept_ and coef_ return the last estimated alpha
-            if self.fit_intercept:
-                self.intercept_path_, self.coef_path_ = _unstandardize(
-                    self.col_means_, self.col_stds_, coef[:, 0], coef[:, 1:]
-                )
-                self.intercept_ = self.intercept_path_[-1]  # type: ignore
-                self.coef_ = self.coef_path_[-1]
-            else:
-                # set intercept to zero as the other linear models do
-                self.intercept_path_, self.coef_path_ = _unstandardize(
-                    self.col_means_, self.col_stds_, np.zeros(coef.shape[0]), coef
-                )
-                self.intercept_ = 0.0
-                self.coef_ = self.coef_path_[-1]
+    # P1 and P2 are now for sure copies
+    P1 = alpha * l1_ratio * P1  # type: ignore
+
+    return typing.cast(np.ndarray, P1).astype(dtype)
+
+
+def setup_p2(
+    P2: Optional[Union[str, np.ndarray, sparse.spmatrix]],
+    X: Union[tm.MatrixBase, tm.StandardizedMatrix],
+    stype,
+    dtype,
+    alpha: float,
+    l1_ratio: float,
+) -> Union[np.ndarray, sparse.spmatrix]:
+    if not isinstance(X, (tm.MatrixBase, tm.StandardizedMatrix)):
+        raise TypeError
+
+    n_features = X.shape[1]
+
+    def _setup_sparse_p2(P2):
+        return (sparse.dia_matrix((P2, 0), shape=(n_features, n_features))).tocsc()
+
+    if isinstance(P2, str):
+        if P2 != "identity":
+            raise ValueError(f"P2 must be either 'identity' or an array. Got {P2}.")
+        if sparse.issparse(X):  # if X is sparse, make P2 sparse, too
+            P2 = _setup_sparse_p2(np.ones(n_features, dtype=dtype))
         else:
-            if self.alpha is None:
-                _alpha = 0.0
-            else:
-                _alpha = self.alpha
-            if _alpha > 0 and self.l1_ratio > 0 and self._solver != "irls-cd":
-                raise ValueError(
-                    f"The chosen solver (solver={self._solver}) can't deal "
-                    "with L1 penalties, which are included with "
-                    f"(alpha={_alpha}) and (l1_ratio={self.l1_ratio})."
-                )
-            coef = self._solve(
-                X=X,
-                y=y,
-                sample_weight=sample_weight,
-                P2=P2_no_alpha * _alpha,
-                P1=P1_no_alpha * _alpha,
-                coef=coef,
-                offset=offset,
-                lower_bounds=lower_bounds,
-                upper_bounds=upper_bounds,
-                A_ineq=A_ineq,
-                b_ineq=b_ineq,
-            )
-
-            if self.fit_intercept:
-                self.intercept_, self.coef_ = _unstandardize(
-                    self.col_means_, self.col_stds_, coef[0], coef[1:]
-                )
-            else:
-                # set intercept to zero as the other linear models do
-                self.intercept_, self.coef_ = _unstandardize(
-                    self.col_means_, self.col_stds_, 0.0, coef
-                )
-
-        self.covariance_matrix_ = None
-        if store_covariance_matrix:
-            self.covariance_matrix(
-                X=X.unstandardize(),
-                y=y,
-                offset=offset,
-                sample_weight=sample_weight * weights_sum,
-                robust=getattr(self, "robust", True),
-                clusters=clusters,
-                expected_information=getattr(self, "expected_information", False),
-                store_covariance_matrix=True,
-                skip_checks=True,
-            )
-
-        return self
-
-    def _compute_information_criteria(
-        self,
-        X: ShapedArrayLike,
-        y: ShapedArrayLike,
-        sample_weight: Optional[ArrayLike] = None,
-        context: Optional[Mapping[str, Any]] = None,
-    ):
-        """
-        Computes and stores the model's degrees of freedom, the 'aic', 'aicc'
-        and 'bic' information criteria.
-
-        The model's degrees of freedom are used to calculate the effective
-        number of parameters. This uses the claim by [2] and [3] that, for
-        L1 regularisation, the number of non-zero parameters in the trained model
-        is an unbiased approximator of the degrees of freedom of the model. Note
-        that this might not hold true for L2 regularisation and thus we raise a
-        warning for this case.
-
-        References
-        ----------
-        [1] Burnham KP, Anderson KR (2002). Model Selection and Multimodel
-        Inference; Springer New York.
-        [2] Zou, H., Hastie, T. and Tibshirani, R., (2007). On the “degrees of
-        freedom” of the lasso; The Annals of Statistics.
-        [3] Park, M.Y., 2006. Generalized linear models with regularization;
-        Stanford Universty.
-        """
-        if not hasattr(self.family_instance, "log_likelihood"):
-            raise NotImplementedError(
-                "The family instance does not define a `log_likelihood` method, so "
-                "information criteria cannot be computed. Compatible families include "
-                "the binomial, negative binomial and Tweedie (power<=2 or power=3)."
-            )
-
-        ddof = np.sum(np.abs(self.coef_) > np.finfo(self.coef_.dtype).eps)  # type: ignore
-        k_params = ddof + self.fit_intercept
-        nobs = X.shape[0]
-
-        if nobs != self._num_obs:
-            raise ValueError(
-                "The same dataset that was used for training should also be used for "
-                "the computation of information criteria."
-            )
-
-        mu = self.predict(X, context=context)
-        ll = self.family_instance.log_likelihood(y, mu, sample_weight=sample_weight)
-
-        aic = -2 * ll + 2 * k_params
-        bic = -2 * ll + np.log(nobs) * k_params
-
-        if nobs > k_params + 1:
-            aicc = aic + 2 * k_params * (k_params + 1) / (nobs - k_params - 1)
+            P2 = np.ones(n_features, dtype=dtype)
+    elif P2 is None:
+        if sparse.issparse(X):  # if X is sparse, make P2 sparse, too
+            P2 = _setup_sparse_p2(np.ones(n_features, dtype=dtype))
         else:
-            aicc = None
-
-        self._info_criteria = {"aic": aic, "aicc": aicc, "bic": bic}
-
-        return True
-
-    def aic(
-        self,
-        X: ArrayLike,
-        y: ArrayLike,
-        sample_weight: Optional[ArrayLike] = None,
-        *,
-        context: Optional[Union[int, Mapping[str, Any]]] = None,
-    ):
-        """
-        Akaike's information criteria. Computed as:
-        :math:`-2\\log\\hat{\\mathcal{L}} + 2\\hat{k}` where
-        :math:`\\hat{\\mathcal{L}}` is the maximum likelihood estimate of the
-        model, and :math:`\\hat{k}` is the effective number of parameters. See
-        `_compute_information_criteria` for more information on the computation
-        of :math:`\\hat{k}`.
-
-        Parameters
-        ----------
-        X : {array-like, sparse matrix}, shape (n_samples, n_features)
-            Same data as used in 'fit'
-
-        y : array-like, shape (n_samples,)
-            Same data as used in 'fit'
-
-        sample_weight : array-like, shape (n_samples,), optional (default=None)
-            Same data as used in 'fit'
-
-        context : Optional[Union[int, Mapping[str, Any]]], default=None
-            The context to add to the evaluation context of the formula with,
-            e.g., custom transforms. If an integer, the context is taken from
-            the stack frame of the caller at the given depth. Otherwise, a
-            mapping from variable names to values is expected. By default,
-            no context is added. Set ``context=0`` to make the calling scope
-            available.
-        """
-        captured_context = capture_context(
-            context + 1 if isinstance(context, int) else context
+            P2 = np.ones(n_features, dtype=dtype)
+    else:
+        P2 = skl.utils.check_array(
+            P2, copy=True, accept_sparse=stype, dtype=dtype, ensure_2d=False
         )
-        return self._get_info_criteria(
-            "aic", X, y, sample_weight, context=captured_context
-        )
-
-    def aicc(
-        self,
-        X: ArrayLike,
-        y: ArrayLike,
-        sample_weight: Optional[ArrayLike] = None,
-        *,
-        context: Optional[Union[int, Mapping[str, Any]]] = None,
-    ):
-        """
-        Second-order Akaike's information criteria (or small sample AIC).
-        Computed as:
-        :math:`-2\\log\\hat{\\mathcal{L}} + 2\\hat{k} + \\frac{2k(k+1)}{n-k-1}`
-        where :math:`\\hat{\\mathcal{L}}` is the maximum likelihood estimate of
-        the model, :math:`n` is the number of training instances, and
-        :math:`\\hat{k}` is the effective number of parameters. See
-        `_compute_information_criteria` for more information on the computation
-        of :math:`\\hat{k}`.
-
-        Parameters
-        ----------
-        X : {array-like, sparse matrix}, shape (n_samples, n_features)
-            Same data as used in 'fit'
-
-        y : array-like, shape (n_samples,)
-            Same data as used in 'fit'
-
-        sample_weight : array-like, shape (n_samples,), optional (default=None)
-             Same data as used in 'fit'
-
-        context : Optional[Union[int, Mapping[str, Any]]], default=None
-            The context to add to the evaluation context of the formula with,
-            e.g., custom transforms. If an integer, the context is taken from
-            the stack frame of the caller at the given depth. Otherwise, a
-            mapping from variable names to values is expected. By default,
-            no context is added. Set ``context=0`` to make the calling scope
-            available.
-        """
-        captured_context = capture_context(
-            context + 1 if isinstance(context, int) else context
-        )
-        aicc = self._get_info_criteria(
-            "aicc", X, y, sample_weight, context=captured_context
-        )
-        if not aicc:
+        P2 = typing.cast(np.ndarray, P2)
+        if P2.ndim == 1:
+            P2 = np.asarray(P2)
+            if P2.shape[0] != n_features:
+                raise ValueError(
+                    "P2 should be a 1d array of shape X.shape[1] (either before or "
+                    "after categorical expansion); "
+                    f"got (P2.shape={P2.shape})."
+                )
+            if sparse.issparse(X):
+                P2 = _setup_sparse_p2(P2)
+        elif P2.ndim == 2 and P2.shape[0] == P2.shape[1] and P2.shape[0] == n_features:
+            if sparse.issparse(X):
+                P2 = sparse.csc_matrix(P2)
+        else:
             raise ValueError(
-                "Model degrees of freedom should be more than training datapoints."
-            )
-        return aicc
-
-    def bic(
-        self,
-        X: ArrayLike,
-        y: ArrayLike,
-        sample_weight: Optional[ArrayLike] = None,
-        *,
-        context: Optional[Union[int, Mapping[str, Any]]] = None,
-    ):
-        """
-        Bayesian information criterion. Computed as:
-        :math:`-2\\log\\hat{\\mathcal{L}} + k\\log(n)` where
-        :math:`\\hat{\\mathcal{L}}` is the maximum likelihood estimate of the
-        model, :math:`n` is the number of training instances, and
-        :math:`\\hat{k}` is the effective number of parameters. See
-        `_compute_information_criteria` for more information on the computation
-        of :math:`\\hat{k}`.
-
-        Parameters
-        ----------
-        X : {array-like, sparse matrix}, shape (n_samples, n_features)
-            Same data as used in 'fit'
-
-        y : array-like, shape (n_samples,)
-            Same data as used in 'fit'
-
-        sample_weight : array-like, shape (n_samples,), optional (default=None)
-             Same data as used in 'fit'
-
-        context : Optional[Union[int, Mapping[str, Any]]], default=None
-            The context to add to the evaluation context of the formula with,
-            e.g., custom transforms. If an integer, the context is taken from
-            the stack frame of the caller at the given depth. Otherwise, a
-            mapping from variable names to values is expected. By default,
-            no context is added. Set ``context=0`` to make the calling scope
-            available.
-        """
-        captured_context = capture_context(
-            context + 1 if isinstance(context, int) else context
-        )
-        return self._get_info_criteria(
-            "bic", X, y, sample_weight, context=captured_context
-        )
-
-    def _get_info_criteria(
-        self,
-        crit: str,
-        X: ArrayLike,
-        y: ArrayLike,
-        sample_weight: Optional[ArrayLike] = None,
-        context: Optional[Mapping[str, Any]] = None,
-    ):
-        check_is_fitted(self, "coef_")
-
-        if not hasattr(self, "_info_criteria"):
-            self._compute_information_criteria(X, y, sample_weight, context=context)
-
-        if (
-            self.alpha is None or (self.alpha is not None and self.alpha > 0)
-        ) and self.l1_ratio < 1.0:
-            warnings.warn(
-                "There is no general definition for the model's degrees of "
-                + f"freedom under L2 (ridge) regularisation. The {crit} "
-                + "might not be well defined in these cases."
+                "P2 must be either None or an array of shape (n_features, n_features) "
+                f"with n_features=X.shape[1]; got (P2.shape={P2.shape}); "
+                f"needed ({n_features}, {n_features})."
             )
 
-        return self._info_criteria[crit]
+    # P1 and P2 are now for sure copies
+    P2 = alpha * (1 - l1_ratio) * P2
+    # one only ever needs the symmetrized L2 penalty matrix 1/2 (P2 + P2')
+    # reason: w' P2 w = (w' P2 w)', i.e. it is symmetric
+    if P2.ndim == 2:
+        if sparse.issparse(P2):
+            if sparse.isspmatrix_csc(P2):
+                P2 = 0.5 * (P2 + P2.transpose()).tocsc()
+            else:
+                P2 = 0.5 * (P2 + P2.transpose()).tocsr()
+        else:
+            P2 = 0.5 * (P2 + P2.T)
+    return P2
+
+
+def _group_sum(groups: np.ndarray, data: tm.MatrixBase):
+    """Sum over groups."""
+    ngroups = len(np.unique(groups))
+    out = np.empty((ngroups, data.shape[1]))
+    eye_n = sps.eye(ngroups, format="csc")[:, groups]
+    for i in range(data.shape[1]):
+        out[:, i] = safe_toarray(eye_n @ data.getcol(i).unpack()).ravel()
+    return out
