+++ tests/glm/test_distribution.py
@@ -3,6 +3,7 @@ import pytest
 import scipy as sp
 import tabmat as tm
 
+from glum._algebra import _safe_sandwich_dot
 from glum._distribution import (
     BinomialDistribution,
     ExponentialDispersionModel,
@@ -14,9 +15,9 @@ from glum._distribution import (
     PoissonDistribution,
     TweedieDistribution,
 )
-from glum._glm import GeneralizedLinearRegressor, get_family
+from glum._glm import get_family
+from glum._glm_regressor import GeneralizedLinearRegressor
 from glum._link import IdentityLink, LogitLink, LogLink, TweedieLink
-from glum._util import _safe_sandwich_dot
 
 
 @pytest.mark.parametrize(
