+++ tests/glm/test_glm_cv.py
@@ -1,10 +1,9 @@
 import numpy as np
 import pandas as pd
 import pytest
+import sklearn as skl
 import tabmat as tm
-from scipy import sparse as sparse
-from sklearn.datasets import make_regression
-from sklearn.linear_model import ElasticNetCV, RidgeCV
+from scipy import sparse
 
 from glum import GeneralizedLinearRegressorCV
 
@@ -37,7 +36,7 @@ def test_normal_elastic_net_comparison(l1_ratio, fit_intercept, convert_x_fn):
     tol = 1e-9
 
     n_predict = 10
-    X, y, _ = make_regression(
+    X, y, _ = skl.datasets.make_regression(
         n_samples=n_samples + n_predict,
         n_features=n_features,
         n_informative=n_features - 2,
@@ -51,7 +50,7 @@ def test_normal_elastic_net_comparison(l1_ratio, fit_intercept, convert_x_fn):
 
     x_arr = X if isinstance(X, np.ndarray) else X.toarray()
     t_arr = T if isinstance(T, np.ndarray) else T.toarray()
-    elastic_net = ElasticNetCV(
+    elastic_net = skl.linear_model.ElasticNetCV(
         l1_ratio=l1_ratio,
         n_alphas=n_alphas,
         fit_intercept=fit_intercept,
@@ -96,7 +95,7 @@ def test_normal_ridge_comparison(fit_intercept):
     alphas = [1e-4]
 
     n_predict = 10
-    X, y, coef = make_regression(
+    X, y, coef = skl.datasets.make_regression(
         n_samples=n_samples + n_predict,
         n_features=n_features,
         n_informative=n_features - 2,
@@ -107,7 +106,9 @@ def test_normal_ridge_comparison(fit_intercept):
     y = y[0:n_samples]
     X, T = X[0:n_samples], X[n_samples:]
 
-    ridge = RidgeCV(fit_intercept=fit_intercept, cv=5, alphas=alphas).fit(X, y)
+    ridge = skl.linear_model.RidgeCV(
+        fit_intercept=fit_intercept, cv=5, alphas=alphas
+    ).fit(X, y)
     el_pred = ridge.predict(T)
 
     glm = GeneralizedLinearRegressorCV(
@@ -124,6 +125,72 @@ def test_normal_ridge_comparison(fit_intercept):
     np.testing.assert_allclose(glm_pred, el_pred, atol=4e-6)
     np.testing.assert_allclose(glm.intercept_, ridge.intercept_, atol=4e-7)
     np.testing.assert_allclose(glm.coef_, ridge.coef_, atol=3e-6)
+
+
+# TODO: different distributions
+# Specify rtol since some are more accurate than others
+@pytest.mark.parametrize(
+    "params",
+    [
+        {"solver": "irls-ls", "rtol": 1e-6},
+        {"solver": "lbfgs", "rtol": 2e-4},
+        {"solver": "trust-constr", "rtol": 2e-4},
+        {"solver": "irls-cd", "selection": "cyclic", "rtol": 2e-5},
+        {"solver": "irls-cd", "selection": "random", "rtol": 6e-5},
+    ],
+    ids=lambda params: ", ".join(f"{key}={val}" for key, val in params.items()),
+)
+@pytest.mark.parametrize("use_offset", [False, True])
+def test_solver_equivalence_cv(params, use_offset):
+    n_alphas = 3
+    n_samples = 100
+    n_features = 10
+    gradient_tol = 1e-5
+
+    X, y = skl.datasets.make_regression(
+        n_samples=n_samples, n_features=n_features, random_state=2
+    )
+
+    if use_offset:
+        np.random.seed(0)
+        offset = np.random.random(len(y))
+    else:
+        offset = None
+
+    est_ref = GeneralizedLinearRegressorCV(
+        random_state=2,
+        n_alphas=n_alphas,
+        gradient_tol=gradient_tol,
+        min_alpha_ratio=1e-3,
+    )
+    est_ref.fit(X, y, offset=offset)
+
+    est_2 = (
+        GeneralizedLinearRegressorCV(
+            n_alphas=n_alphas,
+            max_iter=1000,
+            gradient_tol=gradient_tol,
+            **{k: v for k, v in params.items() if k != "rtol"},
+            min_alpha_ratio=1e-3,
+        )
+        .set_params(random_state=2)
+        .fit(X, y, offset=offset)
+    )
+
+    def _assert_all_close(x, y):
+        return np.testing.assert_allclose(x, y, rtol=params["rtol"], atol=1e-7)
+
+    _assert_all_close(est_2.alphas_, est_ref.alphas_)
+    _assert_all_close(est_2.alpha_, est_ref.alpha_)
+    _assert_all_close(est_2.l1_ratio_, est_ref.l1_ratio_)
+    _assert_all_close(est_2.coef_path_, est_ref.coef_path_)
+    _assert_all_close(est_2.deviance_path_, est_ref.deviance_path_)
+    _assert_all_close(est_2.intercept_, est_ref.intercept_)
+    _assert_all_close(est_2.coef_, est_ref.coef_)
+    _assert_all_close(
+        skl.metrics.mean_absolute_error(est_2.predict(X), y),
+        skl.metrics.mean_absolute_error(est_ref.predict(X), y),
+    )
 
 
 def test_formula():
