+++ tests/glm/test_utils.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from glum._util import _add_missing_categories, _align_df_categories
+from glum._util import add_missing_categories, align_df_categories
 
 
 @pytest.fixture()
@@ -21,7 +21,7 @@ def df():
     )
 
 
-def test_align_df_categories_numeric(df):
+def testalign_df_categories_numeric(df):
     dtypes = {column: np.float64 for column in df}
     has_missing_category = {column: False for column in df}
     missing_method = "fail"
@@ -40,11 +40,11 @@ def test_align_df_categories_numeric(df):
     )
 
     pd.testing.assert_frame_equal(
-        _align_df_categories(df, dtypes, has_missing_category, missing_method), expected
+        align_df_categories(df, dtypes, has_missing_category, missing_method), expected
     )
 
 
-def test_align_df_categories_categorical(df):
+def testalign_df_categories_categorical(df):
     df = df[["x5", "x6", "x7", "x8"]]
     dtypes = {column: pd.CategoricalDtype(["a", "b"]) for column in df}
     has_missing_category = {column: False for column in df}
@@ -61,12 +61,12 @@ def test_align_df_categories_categorical(df):
     )
 
     pd.testing.assert_frame_equal(
-        _align_df_categories(df, dtypes, has_missing_category, missing_method),
+        align_df_categories(df, dtypes, has_missing_category, missing_method),
         expected,
     )
 
 
-def test_align_df_categories_excess_columns(df):
+def testalign_df_categories_excess_columns(df):
     dtypes = {"x1": np.float64}
     has_missing_category = {column: False for column in df}
     missing_method = "fail"
@@ -85,11 +85,11 @@ def test_align_df_categories_excess_columns(df):
     )
 
     pd.testing.assert_frame_equal(
-        _align_df_categories(df, dtypes, has_missing_category, missing_method), expected
+        align_df_categories(df, dtypes, has_missing_category, missing_method), expected
     )
 
 
-def test_align_df_categories_missing_columns(df):
+def testalign_df_categories_missing_columns(df):
     dtypes = {"x0": np.float64}
     has_missing_category = {column: False for column in df}
     missing_method = "fail"
@@ -108,12 +108,12 @@ def test_align_df_categories_missing_columns(df):
     )
 
     pd.testing.assert_frame_equal(
-        _align_df_categories(df, dtypes, has_missing_category, missing_method), expected
+        align_df_categories(df, dtypes, has_missing_category, missing_method), expected
     )
 
 
 @pytest.mark.parametrize("has_missings", [False, True])
-def test_align_df_categories_convert(df, has_missings):
+def testalign_df_categories_convert(df, has_missings):
     df = df[["x5", "x6", "x7", "x8"]]
     dtypes = {column: pd.CategoricalDtype(["a", "b"]) for column in df}
     has_missing_category = {column: has_missings for column in df}
@@ -131,7 +131,7 @@ def test_align_df_categories_convert(df, has_missings):
 
     if has_missings:
         pd.testing.assert_frame_equal(
-            _align_df_categories(
+            align_df_categories(
                 df[["x5", "x6", "x7", "x8"]],
                 dtypes,
                 has_missing_category,
@@ -141,7 +141,7 @@ def test_align_df_categories_convert(df, has_missings):
         )
     else:
         with pytest.raises(ValueError, match="contains unseen categories"):
-            _align_df_categories(
+            align_df_categories(
                 df[["x5", "x6", "x7", "x8"]],
                 dtypes,
                 has_missing_category,
@@ -149,13 +149,13 @@ def test_align_df_categories_convert(df, has_missings):
             )
 
 
-def test_align_df_categories_raise_on_unseen(df):
+def testalign_df_categories_raise_on_unseen(df):
     dtypes = {column: pd.CategoricalDtype(["a", "b"]) for column in df}
     has_missing_category = {column: False for column in df}
     missing_method = "fail"
 
     with pytest.raises(ValueError, match="contains unseen categories"):
-        _align_df_categories(
+        align_df_categories(
             df,
             dtypes,
             has_missing_category,
@@ -163,9 +163,9 @@ def test_align_df_categories_raise_on_unseen(df):
         )
 
 
-def test_align_df_categories_not_df():
+def testalign_df_categories_not_df():
     with pytest.raises(TypeError):
-        _align_df_categories(np.array([[0], [1]]), {"x0": np.float64}, {}, "fail")
+        align_df_categories(np.array([[0], [1]]), {"x0": np.float64}, {}, "fail")
 
 
 @pytest.fixture()
@@ -180,7 +180,7 @@ def df_na():
     )
 
 
-def test_add_missing_categories(df_na):
+def testadd_missing_categories(df_na):
     categorical_format = "{name}[{category}]"
     cat_missing_name = "(M)"
     dtypes = df_na.dtypes
@@ -206,7 +206,7 @@ def test_add_missing_categories(df_na):
     )
 
     pd.testing.assert_frame_equal(
-        _add_missing_categories(
+        add_missing_categories(
             df=df_na,
             dtypes=dtypes,
             feature_names=feature_names,
@@ -238,7 +238,7 @@ def test_raise_on_existing_missing(df_na):
     df.loc[df.cat_na.isna(), "cat_na"] = "(M)"
 
     with pytest.raises(ValueError):
-        _add_missing_categories(
+        add_missing_categories(
             df=df,
             dtypes=dtypes,
             feature_names=feature_names,
